Port proptests
All wallet tests are now ported to the generic framework, and running
for CAPE and Spectrum.
