+++ zerok/zerok_lib/src/lib.rs
@@ -2,10 +2,6 @@
 
 #[cfg(test)]
 #[macro_use]
-extern crate proptest;
-
-#[cfg(test)]
-#[macro_use]
 extern crate quickcheck_macros;
 
 pub mod api;
