+++ zerok/zerok_lib/src/wallet/mod.rs
@@ -2301,1334 +2301,3 @@ impl<'a, L: 'static + Ledger, Backend: 'a + WalletBackend<'a, L> + Send + Sync>
 pub fn new_key_pair() -> UserKeyPair {
     UserKeyPair::generate(&mut ChaChaRng::from_entropy())
 }
-
-#[cfg(any(test, fuzzing))]
-pub mod test_helpers {
-    use super::*;
-    use crate::{
-        api::FromError,
-        node,
-        state::{
-            Block, ElaboratedBlock, ElaboratedTransaction, SetMerkleProof, SetMerkleTree,
-            ValidatorState, VerifierKeySet, MERKLE_HEIGHT,
-        },
-        universal_params::UNIVERSAL_PARAM,
-    };
-    use chrono::Duration;
-    use futures::channel::mpsc as channel;
-    use futures::future;
-    use itertools::izip;
-    use jf_aap::{structs::RecordCommitment, MerkleTree, TransactionVerifyingKey};
-    use key_set::KeySet;
-    use phaselock::traits::state::State;
-    use phaselock::BlockContents;
-    use rand_chacha::rand_core::RngCore;
-    use snafu::ResultExt;
-    use std::iter::once;
-    use std::pin::Pin;
-    use std::sync::Mutex as SyncMutex;
-    use std::time::Instant;
-
-    // This function checks probabilistic equality for two wallet states, comparing hashes for
-    // fields that cannot directly be compared for equality. It is sufficient for tests that want
-    // to compare wallet states (like round-trip serialization tests) but since it is deterministic,
-    // we shouldn't make it into a PartialEq instance.
-    pub fn assert_wallet_states_eq(w1: &WalletState, w2: &WalletState) {
-        assert_eq!(w1.txn_state.now, w2.txn_state.now);
-        assert_eq!(
-            w1.txn_state.validator.commit(),
-            w2.txn_state.validator.commit()
-        );
-        assert_eq!(w1.proving_keys, w2.proving_keys);
-        assert_eq!(w1.txn_state.records, w2.txn_state.records);
-        assert_eq!(w1.key_state, w2.key_state);
-        assert_eq!(w1.auditable_assets, w2.auditable_assets);
-        assert_eq!(
-            w1.audit_keys.keys().collect::<Vec<_>>(),
-            w2.audit_keys.keys().collect::<Vec<_>>()
-        );
-        assert_eq!(
-            w1.freeze_keys.keys().collect::<Vec<_>>(),
-            w2.freeze_keys.keys().collect::<Vec<_>>()
-        );
-        assert_eq!(
-            w1.txn_state.nullifiers.hash(),
-            w2.txn_state.nullifiers.hash()
-        );
-        assert_eq!(
-            w1.txn_state.record_mt.commitment(),
-            w2.txn_state.record_mt.commitment()
-        );
-        assert_eq!(w1.defined_assets, w2.defined_assets);
-        assert_eq!(w1.txn_state.transactions, w2.txn_state.transactions);
-        assert_eq!(w1.key_scans, w2.key_scans);
-    }
-
-    #[derive(Clone, Debug)]
-    pub struct TxnHistoryWithTimeTolerantEq(pub TransactionHistoryEntry<AAPLedger>);
-
-    impl PartialEq<Self> for TxnHistoryWithTimeTolerantEq {
-        fn eq(&self, other: &Self) -> bool {
-            let time_tolerance = Duration::minutes(5);
-            let times_eq = if self.0.time < other.0.time {
-                other.0.time - self.0.time < time_tolerance
-            } else {
-                self.0.time - other.0.time < time_tolerance
-            };
-            times_eq
-                && self.0.asset == other.0.asset
-                && self.0.kind == other.0.kind
-                && self.0.receivers == other.0.receivers
-                && self.0.receipt == other.0.receipt
-        }
-    }
-
-    #[derive(Clone, Debug, Default)]
-    pub struct MockWalletStorage<'a> {
-        committed: Option<WalletState<'a>>,
-        working: Option<WalletState<'a>>,
-        txn_history: Vec<TransactionHistoryEntry<AAPLedger>>,
-    }
-
-    impl<'a> MockWalletStorage<'a> {
-        pub fn store_initial_state(&mut self, state: &WalletState<'a>) {
-            self.committed = Some(state.clone());
-            self.working = Some(state.clone());
-        }
-    }
-
-    #[async_trait]
-    impl<'a> WalletStorage<'a, AAPLedger> for MockWalletStorage<'a> {
-        fn exists(&self) -> bool {
-            self.committed.is_some()
-        }
-
-        async fn load(&mut self) -> Result<WalletState<'a>, WalletError> {
-            Ok(self.committed.as_ref().unwrap().clone())
-        }
-
-        async fn store_snapshot(&mut self, state: &WalletState<'a>) -> Result<(), WalletError> {
-            if let Some(working) = &mut self.working {
-                working.txn_state = state.txn_state.clone();
-                working.key_scans = state.key_scans.clone();
-                working.key_state = state.key_state.clone();
-            }
-            Ok(())
-        }
-
-        async fn store_auditable_asset(
-            &mut self,
-            asset: &AssetDefinition,
-        ) -> Result<(), WalletError> {
-            if let Some(working) = &mut self.working {
-                working.auditable_assets.insert(asset.code, asset.clone());
-            }
-            Ok(())
-        }
-
-        async fn store_key(&mut self, key: &RoleKeyPair) -> Result<(), WalletError> {
-            if let Some(working) = &mut self.working {
-                match key {
-                    RoleKeyPair::Auditor(key) => {
-                        working.audit_keys.insert(key.pub_key(), key.clone());
-                    }
-                    RoleKeyPair::Freezer(key) => {
-                        working.freeze_keys.insert(key.pub_key(), key.clone());
-                    }
-                    RoleKeyPair::User(key) => {
-                        working.user_keys.insert(key.address(), key.clone());
-                    }
-                }
-            }
-            Ok(())
-        }
-
-        async fn store_defined_asset(
-            &mut self,
-            asset: &AssetDefinition,
-            seed: AssetCodeSeed,
-            desc: &[u8],
-        ) -> Result<(), WalletError> {
-            if let Some(working) = &mut self.working {
-                working
-                    .defined_assets
-                    .insert(asset.code, (asset.clone(), seed, desc.to_vec()));
-            }
-            Ok(())
-        }
-
-        async fn store_transaction(
-            &mut self,
-            txn: TransactionHistoryEntry<AAPLedger>,
-        ) -> Result<(), WalletError> {
-            self.txn_history.push(txn);
-            Ok(())
-        }
-
-        async fn transaction_history(
-            &mut self,
-        ) -> Result<Vec<TransactionHistoryEntry<AAPLedger>>, WalletError> {
-            Ok(self.txn_history.clone())
-        }
-
-        async fn commit(&mut self) {
-            self.committed = self.working.clone();
-        }
-
-        async fn revert(&mut self) {
-            self.working = self.committed.clone();
-        }
-    }
-
-    pub struct MockLedger<'a> {
-        pub validator: ValidatorState,
-        nullifiers: SetMerkleTree,
-        records: MerkleTree,
-        subscribers: Vec<channel::UnboundedSender<LedgerEvent>>,
-        current_block: ElaboratedBlock,
-        committed_blocks: Vec<(ElaboratedBlock, Vec<Vec<u64>>)>,
-        block_size: usize,
-        hold_next_transaction: bool,
-        held_transaction: Option<ElaboratedTransaction>,
-        proving_keys: Arc<ProverKeySet<'a, key_set::OrderByOutputs>>,
-        address_map: HashMap<UserAddress, UserPubKey>,
-        events: Vec<LedgerEvent>,
-        storage: Vec<Arc<Mutex<MockWalletStorage<'a>>>>,
-    }
-
-    impl<'a> MockLedger<'a> {
-        pub fn now(&self) -> u64 {
-            self.events.len() as u64
-        }
-
-        fn generate_event(&mut self, e: LedgerEvent) {
-            println!(
-                "generating event {}: {}",
-                self.events.len(),
-                match &e {
-                    LedgerEvent::Commit { .. } => "Commit",
-                    LedgerEvent::Reject { .. } => "Reject",
-                    LedgerEvent::Memos { .. } => "Memos",
-                }
-            );
-            self.events.push(e.clone());
-            self.subscribers = std::mem::take(&mut self.subscribers)
-                .into_iter()
-                .filter_map(|mut s| {
-                    if s.start_send(e.clone()).is_ok() {
-                        // Errors indicate that the subscriber has disconnected, so we only want to
-                        // retain the subscriber if the send is successful.
-                        Some(s)
-                    } else {
-                        None
-                    }
-                })
-                .collect();
-        }
-
-        pub fn flush(&mut self) {
-            if self.current_block.block.0.is_empty() {
-                return;
-            }
-
-            let block = std::mem::replace(&mut self.current_block, self.validator.next_block());
-            match self.validator.validate_and_apply(
-                self.validator.prev_commit_time + 1,
-                block.block.clone(),
-                block.proofs.clone(),
-            ) {
-                Ok(mut uids) => {
-                    // Add nullifiers
-                    for txn in &block.block.0 {
-                        for nullifier in txn.nullifiers() {
-                            self.nullifiers.insert(nullifier);
-                        }
-                        for record in txn.output_commitments() {
-                            self.records.push(record.to_field_element())
-                        }
-                    }
-
-                    // Broadcast the new block
-                    self.generate_event(LedgerEvent::Commit {
-                        block: block.clone(),
-                        block_id: self.committed_blocks.len() as u64,
-                        state_comm: self.validator.commit(),
-                    });
-
-                    // Store the block in the history
-                    let mut block_uids = vec![];
-                    for txn in block.block.0.iter() {
-                        let mut this_txn_uids = uids;
-                        uids = this_txn_uids.split_off(txn.output_len());
-                        assert_eq!(this_txn_uids.len(), txn.output_len());
-                        block_uids.push(this_txn_uids);
-                    }
-                    self.committed_blocks.push((block, block_uids));
-                }
-                Err(error) => self.generate_event(LedgerEvent::Reject { block, error }),
-            }
-        }
-
-        pub fn hold_next_transaction(&mut self) {
-            self.hold_next_transaction = true;
-        }
-
-        pub fn release_held_transaction(&mut self) -> Option<ElaboratedTransaction> {
-            if let Some(txn) = self.held_transaction.take() {
-                self.submit(txn.clone());
-                Some(txn)
-            } else {
-                None
-            }
-        }
-
-        pub fn submit(&mut self, txn: ElaboratedTransaction) {
-            if self.hold_next_transaction {
-                self.held_transaction = Some(txn);
-                self.hold_next_transaction = false;
-                return;
-            }
-
-            match self.current_block.add_transaction_raw(&txn) {
-                Ok(block) => {
-                    self.current_block = block;
-                    // self.current_memos.push(memos);
-                    if self.current_block.block.0.len() >= self.block_size {
-                        self.flush();
-                    }
-                }
-                Err(error) => {
-                    let rejected = ElaboratedBlock {
-                        block: Block(vec![txn.txn]),
-                        proofs: vec![txn.proofs],
-                    };
-                    self.generate_event(LedgerEvent::<AAPLedger>::Reject {
-                        block: rejected,
-                        error,
-                    });
-                }
-            }
-        }
-
-        pub fn post_memos(
-            &mut self,
-            block_id: u64,
-            txn_id: u64,
-            memos: Vec<ReceiverMemo>,
-            sig: Signature,
-        ) -> Result<(), WalletError> {
-            let (block, block_uids) = &self.committed_blocks[block_id as usize];
-            let txn = &block.block.0[txn_id as usize];
-            let comms = txn.output_commitments();
-            let uids = block_uids[txn_id as usize].clone();
-
-            txn.verify_receiver_memos_signature(&memos, &sig)
-                .context(CryptoError)?;
-
-            let merkle_paths = uids
-                .iter()
-                .map(|uid| {
-                    self.records
-                        .get_leaf(*uid)
-                        .expect_ok()
-                        .map(|(_, proof)| (proof.leaf.0, proof.path))
-                        .unwrap()
-                        .1
-                })
-                .collect::<Vec<_>>();
-            self.generate_event(LedgerEvent::<AAPLedger>::Memos {
-                outputs: izip!(memos, comms, uids, merkle_paths).collect(),
-                transaction: Some((block_id, txn_id)),
-            });
-
-            Ok(())
-        }
-    }
-
-    pub async fn sync<'a>(
-        ledger: &Arc<SyncMutex<MockLedger<'a>>>,
-        wallets: &[(
-            Wallet<'a, impl 'a + WalletBackend<'a, AAPLedger> + Send + Sync>,
-            UserAddress,
-        )],
-    ) {
-        let t = {
-            let mut ledger = ledger.lock().unwrap();
-            ledger.flush();
-            if let Some(LedgerEvent::Commit { block, .. }) = ledger.events.last() {
-                // If the last event is a Commit, wait until all of the senders from the block
-                // receive the Commit event and post the receiver memos, generating new Memos events.
-                ledger.now() + (block.block.0.len() as u64)
-            } else {
-                ledger.now()
-            }
-        };
-        sync_with(wallets, t).await;
-
-        // Since we're syncing with the time stamp from the most recent event, the wallets should
-        // be in a stable state once they have processed up to that event. Check that each wallet
-        // has persisted all of its in-memory state at this point.
-        let ledger = ledger.lock().unwrap();
-        for ((wallet, _), storage) in wallets.iter().zip(&ledger.storage) {
-            let WalletSharedState { state, .. } = &*wallet.mutex.lock().await;
-            assert_wallet_states_eq(state, storage.lock().await.committed.as_ref().unwrap());
-        }
-    }
-
-    pub async fn sync_with<'a>(
-        wallets: &[(
-            Wallet<'a, impl 'a + WalletBackend<'a, AAPLedger> + Send + Sync>,
-            UserAddress,
-        )],
-        t: u64,
-    ) {
-        println!("waiting for sync point {}", t);
-        future::join_all(wallets.iter().map(|(wallet, _)| wallet.sync(t))).await;
-    }
-
-    #[derive(Clone)]
-    pub struct MockWalletBackend<'a> {
-        key_pair: UserKeyPair,
-        ledger: Arc<SyncMutex<MockLedger<'a>>>,
-        initial_grants: Vec<(RecordOpening, u64)>,
-        seed: [u8; 32],
-        storage: Arc<Mutex<MockWalletStorage<'a>>>,
-    }
-
-    #[async_trait]
-    impl<'a> WalletBackend<'a, AAPLedger> for MockWalletBackend<'a> {
-        type EventStream = Pin<Box<dyn Stream<Item = LedgerEvent> + Send>>;
-        type Storage = MockWalletStorage<'a>;
-
-        async fn storage<'l>(&'l mut self) -> MutexGuard<'l, Self::Storage> {
-            self.storage.lock().await
-        }
-
-        async fn create(&mut self) -> Result<WalletState<'a>, WalletError> {
-            let state = {
-                let ledger = self.ledger.lock().unwrap();
-
-                WalletState {
-                    proving_keys: ledger.proving_keys.clone(),
-                    txn_state: TransactionState {
-                        validator: ledger.validator.clone(),
-
-                        records: {
-                            let mut db: RecordDatabase = Default::default();
-                            for (ro, uid) in self.initial_grants.iter() {
-                                db.insert(ro.clone(), *uid, &self.key_pair);
-                            }
-                            db
-                        },
-                        nullifiers: ledger.nullifiers.clone(),
-                        record_mt: ledger.records.clone(),
-                        merkle_leaf_to_forget: None,
-
-                        now: 0,
-                        transactions: Default::default(),
-                    },
-                    key_state: Default::default(),
-                    key_scans: Default::default(),
-                    auditable_assets: Default::default(),
-                    audit_keys: Default::default(),
-                    freeze_keys: Default::default(),
-                    user_keys: Default::default(),
-                    defined_assets: HashMap::new(),
-                }
-            };
-
-            // Persist the initial state.
-            let mut storage = self.storage().await;
-            storage.committed = Some(state.clone());
-            storage.working = Some(state.clone());
-
-            Ok(state)
-        }
-
-        fn key_stream(&self) -> hd::KeyTree {
-            let mut rng = ChaChaRng::from_seed(self.seed);
-            hd::KeyTree::random(&mut rng).unwrap().0
-        }
-
-        async fn subscribe(&self, starting_at: u64) -> Self::EventStream {
-            let mut ledger = self.ledger.lock().unwrap();
-
-            assert!(
-                starting_at <= ledger.now(),
-                "subscribing from a future state is not supported in the MockWalletBackend"
-            );
-            let past_events = ledger
-                .events
-                .iter()
-                .skip(starting_at as usize)
-                .cloned()
-                .collect::<Vec<_>>();
-
-            let (sender, receiver) = channel::unbounded();
-            ledger.subscribers.push(sender);
-
-            Box::pin(iter(past_events).chain(receiver))
-        }
-
-        async fn get_public_key(&self, address: &UserAddress) -> Result<UserPubKey, WalletError> {
-            let ledger = self.ledger.lock().unwrap();
-            match ledger.address_map.get(address) {
-                Some(key) => Ok(key.clone()),
-                None => Err(WalletError::InvalidAddress {
-                    address: address.clone(),
-                }),
-            }
-        }
-
-        async fn get_nullifier_proof(
-            &self,
-            set: &mut SetMerkleTree,
-            nullifier: Nullifier,
-        ) -> Result<(bool, SetMerkleProof), WalletError> {
-            let ledger = self.ledger.lock().unwrap();
-            if set.hash() == ledger.nullifiers.hash() {
-                Ok(ledger.nullifiers.contains(nullifier).unwrap())
-            } else {
-                Err(WalletError::QueryServiceError {
-                    source: node::QueryServiceError::InvalidNullifierRoot {},
-                })
-            }
-        }
-
-        async fn get_transaction(
-            &self,
-            block_id: u64,
-            txn_id: u64,
-        ) -> Result<ElaboratedTransaction, WalletError> {
-            let ledger = self.ledger.lock().unwrap();
-            let block = &ledger
-                .committed_blocks
-                .get(block_id as usize)
-                .ok_or_else(|| {
-                    WalletError::from_query_service_error(node::QueryServiceError::InvalidBlockId {
-                        index: block_id as usize,
-                        num_blocks: ledger.committed_blocks.len(),
-                    })
-                })?
-                .0;
-
-            if txn_id as usize >= block.block.0.len() {
-                return Err(WalletError::from_query_service_error(
-                    node::QueryServiceError::InvalidTxnId {},
-                ));
-            }
-            let txn = block.block.0[txn_id as usize].clone();
-            let proofs = block.proofs[txn_id as usize].clone();
-            Ok(ElaboratedTransaction { txn, proofs })
-        }
-
-        async fn register_user_key(&mut self, pub_key: &UserPubKey) -> Result<(), WalletError> {
-            let mut ledger = self.ledger.lock().unwrap();
-            ledger
-                .address_map
-                .insert(pub_key.address(), pub_key.clone());
-            Ok(())
-        }
-
-        async fn submit(&mut self, txn: ElaboratedTransaction) -> Result<(), WalletError> {
-            self.ledger.lock().unwrap().submit(txn);
-            Ok(())
-        }
-
-        async fn post_memos(
-            &mut self,
-            block_id: u64,
-            txn_id: u64,
-            memos: Vec<ReceiverMemo>,
-            sig: Signature,
-        ) -> Result<(), WalletError> {
-            self.ledger
-                .lock()
-                .unwrap()
-                .post_memos(block_id, txn_id, memos, sig)
-        }
-    }
-
-    pub async fn create_test_network<'a>(
-        xfr_sizes: &[(usize, usize)],
-        initial_grants: Vec<u64>,
-        now: &mut Instant,
-    ) -> (
-        Arc<SyncMutex<MockLedger<'a>>>,
-        Vec<(Wallet<'a, MockWalletBackend<'a>>, UserAddress)>,
-    ) {
-        let mut rng = ChaChaRng::from_seed([42u8; 32]);
-
-        // Populate the unpruned record merkle tree with an initial record commitment for each
-        // non-zero initial grant. Collect user-specific info (keys and record openings
-        // corresponding to grants) in `users`, which will be used to create the wallets later.
-        let mut record_merkle_tree = MerkleTree::new(MERKLE_HEIGHT).unwrap();
-        let mut users = vec![];
-        for amount in initial_grants {
-            let key = UserKeyPair::generate(&mut rng);
-            if amount > 0 {
-                let ro = RecordOpening::new(
-                    &mut rng,
-                    amount,
-                    AssetDefinition::native(),
-                    key.pub_key(),
-                    FreezeFlag::Unfrozen,
-                );
-                let comm = RecordCommitment::from(&ro);
-                let uid = record_merkle_tree.num_leaves();
-                record_merkle_tree.push(comm.to_field_element());
-                users.push((key, vec![(ro, uid)]));
-            } else {
-                users.push((key, vec![]));
-            }
-        }
-
-        // Create the validator using the ledger state containing the initial grants, computed above.
-        println!(
-            "Generating validator keys: {}s",
-            now.elapsed().as_secs_f32()
-        );
-        *now = Instant::now();
-
-        let mut xfr_prove_keys = vec![];
-        let mut xfr_verif_keys = vec![];
-        for (num_inputs, num_outputs) in xfr_sizes {
-            let (xfr_prove_key, xfr_verif_key, _) = jf_aap::proof::transfer::preprocess(
-                &*UNIVERSAL_PARAM,
-                *num_inputs,
-                *num_outputs,
-                MERKLE_HEIGHT,
-            )
-            .unwrap();
-            xfr_prove_keys.push(xfr_prove_key);
-            xfr_verif_keys.push(TransactionVerifyingKey::Transfer(xfr_verif_key));
-        }
-        let (mint_prove_key, mint_verif_key, _) =
-            jf_aap::proof::mint::preprocess(&*UNIVERSAL_PARAM, MERKLE_HEIGHT).unwrap();
-        let (freeze_prove_key, freeze_verif_key, _) =
-            jf_aap::proof::freeze::preprocess(&*UNIVERSAL_PARAM, 2, MERKLE_HEIGHT).unwrap();
-        let nullifiers: SetMerkleTree = Default::default();
-        let validator = ValidatorState::new(
-            VerifierKeySet {
-                xfr: KeySet::new(xfr_verif_keys.into_iter()).unwrap(),
-                mint: TransactionVerifyingKey::Mint(mint_verif_key),
-                freeze: KeySet::new(
-                    vec![TransactionVerifyingKey::Freeze(freeze_verif_key)].into_iter(),
-                )
-                .unwrap(),
-            },
-            record_merkle_tree.clone(),
-        );
-
-        let comm = validator.commit();
-        println!(
-            "Validator set up with state {:x?}: {}s",
-            comm,
-            now.elapsed().as_secs_f32()
-        );
-
-        let current_block = validator.next_block();
-        let mut storage = Vec::new();
-        for _ in &users {
-            storage.push(Arc::new(Mutex::new(MockWalletStorage::default())));
-        }
-        let ledger = Arc::new(SyncMutex::new(MockLedger {
-            validator,
-            nullifiers,
-            records: record_merkle_tree,
-            subscribers: Vec::new(),
-            current_block,
-            committed_blocks: Vec::new(),
-            block_size: 2,
-            hold_next_transaction: false,
-            held_transaction: None,
-            proving_keys: Arc::new(ProverKeySet {
-                xfr: KeySet::new(xfr_prove_keys.into_iter()).unwrap(),
-                mint: mint_prove_key,
-                freeze: KeySet::new(vec![freeze_prove_key].into_iter()).unwrap(),
-            }),
-            address_map: users
-                .iter()
-                .map(|(key, _)| (key.address(), key.pub_key()))
-                .collect(),
-            events: Vec::new(),
-            storage: storage.clone(),
-        }));
-
-        // Create a wallet for each user based on the validator and the per-user information
-        // computed above.
-        let wallets = iter(users)
-            .zip(iter(storage))
-            .then(|((key_pair, initial_grants), storage)| {
-                let mut rng = ChaChaRng::from_rng(&mut rng).unwrap();
-                let ledger = ledger.clone();
-                async move {
-                    let mut seed = [0u8; 32];
-                    rng.fill_bytes(&mut seed);
-                    let mut wallet = Wallet::new(MockWalletBackend {
-                        ledger,
-                        initial_grants,
-                        seed,
-                        storage,
-                        key_pair: key_pair.clone(),
-                    })
-                    .await
-                    .unwrap();
-                    wallet.add_user_key(key_pair.clone(), 0).await.unwrap();
-                    (wallet, key_pair.address())
-                }
-            })
-            .collect()
-            .await;
-
-        println!("Wallets set up: {}s", now.elapsed().as_secs_f32());
-        *now = Instant::now();
-        (ledger, wallets)
-    }
-
-    /*
-     * This test is very similar to test_two_wallets, but it is parameterized on the number of users,
-     * number of asset types, initial ledger state, and transactions to do, so it can be used with
-     * quickcheck or proptest to do randomized fuzzing.
-     */
-    #[allow(clippy::type_complexity)]
-    pub async fn test_multixfr_wallet(
-        // List of blocks containing (def,key1,key2,amount) transfer specs
-        // An asset def of 0 in a transfer spec or record indicates the native asset type; other
-        // asset types are indexed startin from 1.
-        txs: Vec<Vec<(u8, u8, u8, u64)>>,
-        nkeys: u8,
-        ndefs: u8,
-        // (def,key,amount)
-        init_rec: (u8, u8, u64),
-        init_recs: Vec<(u8, u8, u64)>,
-    ) {
-        println!(
-            "multixfr_wallet test: {} users, {} assets, {} records, {} transfers",
-            nkeys,
-            ndefs,
-            init_recs.len() + 1,
-            txs.iter().flatten().count()
-        );
-        let mut now = Instant::now();
-
-        let xfr_sizes = &[
-            (1, 2), // basic native transfer
-            (2, 2), // basic non-native transfer, or native merge
-            (2, 3), // non-native transfer with change output
-            (3, 2), // non-native merge
-        ];
-        let mut balances = vec![vec![0; ndefs as usize + 1]; nkeys as usize];
-        // `histories` is a map from wallet indices to vectors of blocks of history entries. The
-        // reason for blocking the history entries is that entries corresponding to transactions
-        // that were validated in the same block can be recorded by the wallets in any order.
-        let mut histories = vec![vec![vec![]]; nkeys as usize];
-        let grants =
-            // The issuer (wallet 0) gets 1 coin per initial record, to pay transaction fees while
-            // it mints and distributes the records, and 1 coin per transaction, to pay transaction
-            // fees while minting additional records if test wallets run out of balance during the
-            // test.
-            once((1 + init_recs.len() + txs.iter().flatten().count()) as u64).chain(
-                (0..nkeys)
-                    .map(|i| {
-                        // The remaining wallets (the test wallets) get 1 coin for each transaction
-                        // in which they are the sender, to pay transaction fees, plus...
-                        let txn_fees = txs.iter()
-                            .flatten()
-                            .map(|(_, sender, _, _)| {
-                                if sender % nkeys == i {1} else {0}
-                            })
-                            .sum::<u64>();
-                        balances[i as usize][0] += txn_fees;
-                        txn_fees +
-                        // ...one record for each native asset type initial record that they own,
-                        // plus...
-                        once(&init_rec).chain(&init_recs)
-                            .map(|(def, owner, amount)| {
-                                let def = (def % (ndefs + 1)) as usize;
-                                let owner = (owner % nkeys) as usize;
-                                if def == 0 && owner == (i as usize) {
-                                    balances[owner][def] += amount;
-                                    *amount
-                                } else {
-                                    0
-                                }
-                            })
-                            .sum::<u64>() +
-                        // We want to prevent transfers of the native asset type from failing due to
-                        // insufficient funds, or worse, from dipping into native coins which were
-                        // intended to be used later as transaction fees. Unlike non-native
-                        // transfers, we can't mint more native coins during the test if we find
-                        // that one of the wallets is low on balance. So we give each wallet an
-                        // extra grant of native coins large enough to cover all the native
-                        // transactions it will need to make, when combined with its original grant
-                        // of native coins.
-                        {
-                            let total_txn_amount: u64 = txs.iter()
-                                .flatten()
-                                .map(|(def, sender, _, amount)| {
-                                    if (def % (ndefs + 1)) == 0 && (sender % nkeys) == i {
-                                        *amount
-                                    } else {
-                                        0
-                                    }
-                                })
-                                .sum();
-                            if txn_fees + total_txn_amount > balances[i as usize][0] {
-                                let extra = txn_fees + total_txn_amount - balances[i as usize][0];
-                                balances[i as usize][0] += extra;
-                                extra
-                            } else {
-                                0
-                            }
-                        }
-                    })
-            ).collect();
-
-        let (ledger, mut wallets) = create_test_network(xfr_sizes, grants, &mut now).await;
-        println!(
-            "ceremony complete, minting initial records: {}s",
-            now.elapsed().as_secs_f32()
-        );
-        now = Instant::now();
-
-        fn push_history(
-            wallet_ix: usize,
-            histories: &mut [Vec<Vec<TransactionHistoryEntry<AAPLedger>>>],
-            entry: TransactionHistoryEntry<AAPLedger>,
-        ) {
-            histories[wallet_ix].last_mut().unwrap().push(entry);
-        }
-        fn close_history_block(histories: &mut [Vec<Vec<TransactionHistoryEntry<AAPLedger>>>]) {
-            for history in histories {
-                history.push(vec![])
-            }
-        }
-
-        // Define all of the test assets and mint initial records.
-        let mut assets = vec![];
-        for i in 0..ndefs {
-            assets.push(
-                wallets[0]
-                    .0
-                    .define_asset(format!("Asset {}", i).as_bytes(), Default::default())
-                    .await
-                    .unwrap(),
-            );
-        }
-        for (asset, owner, amount) in once(init_rec).chain(init_recs) {
-            let asset = (asset % (ndefs + 1)) as usize;
-            if asset == 0 {
-                // can't mint native assets
-                continue;
-            }
-            let minter = wallets[0].1.clone();
-            let address = wallets[(owner % nkeys) as usize + 1].1.clone();
-            balances[(owner % nkeys) as usize][asset] += amount;
-            wallets[0]
-                .0
-                .mint(&minter, 1, &assets[asset - 1].code, amount, address.clone())
-                .await
-                .unwrap();
-            push_history(
-                (owner % nkeys) as usize,
-                &mut histories,
-                TransactionHistoryEntry {
-                    time: Local::now(),
-                    asset: assets[asset - 1].code,
-                    kind: TransactionKind::<AAPLedger>::mint(),
-                    sender: None,
-                    receivers: vec![(address, amount)],
-                    receipt: None,
-                },
-            );
-            sync(&ledger, &wallets).await;
-            close_history_block(&mut histories);
-        }
-
-        println!("assets minted: {}s", now.elapsed().as_secs_f32());
-        now = Instant::now();
-
-        // Check initial balances. This cannot be a closure because rust infers the wrong lifetime
-        // for the references (it tries to use 'a, which is longer than we want to borrow `wallets`
-        // for).
-        async fn check_balances<'a>(
-            wallets: &[(Wallet<'a, MockWalletBackend<'a>>, UserAddress)],
-            balances: &[Vec<u64>],
-            assets: &[AssetDefinition],
-        ) {
-            for (i, balance) in balances.iter().enumerate() {
-                let (wallet, address) = &wallets[i + 1];
-
-                // Check native asset balance.
-                assert_eq!(
-                    wallet.balance(address, &AssetCode::native()).await,
-                    balance[0]
-                );
-                for (j, asset) in assets.iter().enumerate() {
-                    assert_eq!(wallet.balance(address, &asset.code).await, balance[j + 1]);
-                }
-            }
-        }
-        check_balances(&wallets, &balances, &assets).await;
-
-        async fn check_histories<'a>(
-            wallets: &[(Wallet<'a, MockWalletBackend<'a>>, UserAddress)],
-            histories: &[Vec<Vec<TransactionHistoryEntry<AAPLedger>>>],
-        ) {
-            assert_eq!(wallets.len(), histories.len() + 1);
-            for ((wallet, _), history) in wallets.iter().skip(1).zip(histories) {
-                let mut wallet_history = wallet.transaction_history().await.unwrap();
-                assert_eq!(
-                    wallet_history.len(),
-                    history.iter().map(|block| block.len()).sum::<usize>()
-                );
-
-                for block in history {
-                    let remaining = wallet_history.split_off(block.len());
-                    let wallet_block = wallet_history;
-                    wallet_history = remaining;
-
-                    // Compare the blocks, allowing for slight deviations in the timestamps of
-                    // corresponding entries. We compare blocks modulo order by checking that they
-                    // have the same length and that every entry in one is in the other, and vice
-                    // versa.
-                    assert_eq!(wallet_block.len(), block.len());
-                    let wallet_block = wallet_block
-                        .into_iter()
-                        .map(TxnHistoryWithTimeTolerantEq)
-                        .collect::<Vec<_>>();
-                    let block = block
-                        .iter()
-                        .map(|txn| TxnHistoryWithTimeTolerantEq(txn.clone()))
-                        .collect::<Vec<_>>();
-                    for txn in wallet_block.iter() {
-                        assert!(block.contains(txn));
-                    }
-                    for txn in block.iter() {
-                        assert!(wallet_block.contains(txn));
-                    }
-                }
-            }
-        }
-        check_histories(&wallets, &histories).await;
-
-        // Run the test transactions.
-        for (i, block) in txs.iter().enumerate() {
-            println!(
-                "Starting block {}/{}: {}s",
-                i + 1,
-                txs.len(),
-                now.elapsed().as_secs_f32()
-            );
-            now = Instant::now();
-
-            for (j, (asset_ix, sender_ix, receiver_ix, amount)) in block.iter().enumerate() {
-                println!(
-                    "Starting txn {}.{}/{}:{:?}: {}s",
-                    i + 1,
-                    j + 1,
-                    block.len(),
-                    (asset_ix, sender_ix, receiver_ix, amount),
-                    now.elapsed().as_secs_f32()
-                );
-
-                let asset_ix = (asset_ix % (ndefs + 1)) as usize;
-                let sender_ix = (sender_ix % nkeys) as usize;
-                let receiver_ix = (receiver_ix % nkeys) as usize;
-                let native = AssetDefinition::native();
-                let asset = if asset_ix == 0 {
-                    &native
-                } else {
-                    &assets[asset_ix - 1]
-                };
-                let receiver = wallets[receiver_ix + 1].1.clone();
-                let sender_address = wallets[sender_ix + 1].1.clone();
-                let sender_balance = balances[sender_ix][asset_ix];
-
-                let mut amount = if *amount <= sender_balance {
-                    *amount
-                } else if sender_balance > 0 {
-                    // If we don't have enough to make the whole transfer, but we have some,
-                    // transfer half of what we have.
-                    let new_amount = std::cmp::max(sender_balance / 2, 1);
-                    println!(
-                        "decreasing transfer amount due to insufficient balance: {} -> {}: {}s",
-                        *amount,
-                        new_amount,
-                        now.elapsed().as_secs_f32()
-                    );
-                    now = Instant::now();
-                    new_amount
-                } else {
-                    // If we don't have any of this asset type, mint more.
-                    assert_ne!(asset, &AssetDefinition::native());
-                    println!(
-                        "minting {} more of asset {:?}: {}s",
-                        *amount,
-                        &asset.code,
-                        now.elapsed().as_secs_f32()
-                    );
-                    now = Instant::now();
-                    let (minter, minter_address) = &mut wallets[0];
-                    minter
-                        .mint(
-                            minter_address,
-                            1,
-                            &asset.code,
-                            2 * amount,
-                            sender_address.clone(),
-                        )
-                        .await
-                        .unwrap();
-                    sync(&ledger, &wallets).await;
-                    balances[sender_ix][asset_ix] += 2 * amount;
-                    push_history(
-                        sender_ix,
-                        &mut histories,
-                        TransactionHistoryEntry {
-                            time: Local::now(),
-                            asset: asset.code,
-                            kind: TransactionKind::<AAPLedger>::mint(),
-                            sender: None,
-                            receivers: vec![(sender_address.clone(), 2 * amount)],
-                            receipt: None,
-                        },
-                    );
-
-                    println!("asset minted: {}s", now.elapsed().as_secs_f32());
-                    now = Instant::now();
-                    *amount
-                };
-
-                ledger.lock().unwrap().hold_next_transaction();
-                let sender = &mut wallets[sender_ix + 1].0;
-                let receipt = match sender
-                    .transfer(
-                        &sender_address,
-                        &asset.code,
-                        &[(receiver.clone(), amount)],
-                        1,
-                    )
-                    .await
-                {
-                    Ok(receipt) => receipt,
-                    Err(WalletError::TransactionError {
-                        source:
-                            TransactionError::Fragmentation {
-                                suggested_amount, ..
-                            },
-                    }) => {
-                        // Allow fragmentation. Without merge transactions, there's not much we can
-                        // do to prevent it, and merge transactions require multiple transaction
-                        // arities, which requires either dummy records or multiple verifier keys in
-                        // the validator.
-                        if suggested_amount > 0 {
-                            // If the wallet suggested a transaction amount that it _can_ process,
-                            // try again with that amount.
-                            println!(
-                                "decreasing transfer amount due to fragmentation: {} -> {}: {}s",
-                                amount,
-                                suggested_amount,
-                                now.elapsed().as_secs_f32()
-                            );
-                            now = Instant::now();
-
-                            amount = suggested_amount;
-                            sender
-                                .transfer(
-                                    &sender_address,
-                                    &asset.code,
-                                    &[(receiver.clone(), amount)],
-                                    1,
-                                )
-                                .await
-                                .unwrap()
-                        } else {
-                            println!(
-                                "skipping transfer due to fragmentation: {}s",
-                                now.elapsed().as_secs_f32()
-                            );
-                            now = Instant::now();
-                            continue;
-                        }
-                    }
-                    Err(WalletError::TransactionError {
-                        source: TransactionError::InsufficientBalance { .. },
-                    }) => {
-                        // We should always have enough balance to make the transaction, because we
-                        // adjusted the transaction amount (and potentially minted more of the
-                        // asset) above, so that the transaction is covered by our most up-to-date
-                        // balance.
-                        //
-                        // If we fail due to insufficient balance, it is likely because a record we
-                        // need is on hold as part of a previous transaction, and we haven't gotten
-                        // the change yet because the transaction is buffered in a block. The
-                        // transaction should succeed after we flush any pending transactions.
-                        println!("flushing pending blocks to retrieve change");
-                        ledger.lock().unwrap().flush();
-                        sync(&ledger, &wallets).await;
-                        let sender = &mut wallets[sender_ix + 1].0;
-                        sender
-                            .transfer(
-                                &sender_address,
-                                &asset.code,
-                                &[(receiver.clone(), amount)],
-                                1,
-                            )
-                            .await
-                            .unwrap()
-                    }
-                    Err(err) => {
-                        panic!("transaction failed: {:?}", err)
-                    }
-                };
-                println!(
-                    "Generated txn {}.{}/{}: {}s",
-                    i + 1,
-                    j + 1,
-                    block.len(),
-                    now.elapsed().as_secs_f32()
-                );
-                now = Instant::now();
-
-                balances[sender_ix][0] -= 1; // transaction fee
-                balances[sender_ix][asset_ix] -= amount;
-                balances[receiver_ix][asset_ix] += amount;
-
-                push_history(
-                    sender_ix,
-                    &mut histories,
-                    TransactionHistoryEntry {
-                        time: Local::now(),
-                        asset: asset.code,
-                        kind: TransactionKind::<AAPLedger>::send(),
-                        sender: Some(sender_address),
-                        receivers: vec![(receiver.clone(), amount)],
-                        receipt: Some(receipt),
-                    },
-                );
-                if receiver_ix != sender_ix {
-                    push_history(
-                        receiver_ix,
-                        &mut histories,
-                        TransactionHistoryEntry {
-                            time: Local::now(),
-                            asset: asset.code,
-                            kind: TransactionKind::<AAPLedger>::receive(),
-                            sender: None,
-                            receivers: vec![(receiver, amount)],
-                            receipt: None,
-                        },
-                    );
-                }
-
-                ledger.lock().unwrap().release_held_transaction();
-            }
-
-            sync(&ledger, &wallets).await;
-            close_history_block(&mut histories);
-            check_balances(&wallets, &balances, &assets).await;
-            check_histories(&wallets, &histories).await;
-
-            println!(
-                "Finished block {}/{}: {}s",
-                i + 1,
-                block.len(),
-                now.elapsed().as_secs_f32()
-            );
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use async_std::task::block_on;
-    use proptest::collection::vec;
-    use proptest::strategy::Strategy;
-    use test_helpers::*;
-
-    #[async_std::test]
-    async fn test_multixfr_wallet_simple() -> std::io::Result<()> {
-        let alice_grant = (0, 0, 3); // Alice gets 3 of coin 0 to start
-        let bob_grant = (1, 1, 3); // Bob gets 3 of coin 1 to start
-        let txns = vec![vec![
-            (1, 0, 1, 2), // Alice sends 2 of coin 1 to Bob
-            (2, 1, 0, 2), // Bob sends 2 of coin 2 to Alice
-            (1, 1, 0, 1), // Bob sends 1 of coin 1 to Alice
-        ]];
-        test_multixfr_wallet(txns, 2, 2, alice_grant, vec![bob_grant]).await;
-        Ok(())
-    }
-
-    #[async_std::test]
-    async fn test_multixfr_wallet_multi_xfr_block() -> std::io::Result<()> {
-        // Alice and Bob each get 1 native token to start.
-        let alice_grant = (0, 0, 1);
-        let bob_grant = (0, 1, 1);
-        // Alice and Bob make independent transactions, so that the transactions can end up in the
-        // same block.
-        let txns = vec![vec![
-            (0, 0, 1, 1), // Alice sends 1 coin to Bob
-            (0, 1, 0, 1), // Bob sends 1 coin to Alice
-        ]];
-        test_multixfr_wallet(txns, 2, 1, alice_grant, vec![bob_grant]).await;
-        Ok(())
-    }
-
-    #[async_std::test]
-    async fn test_multixfr_wallet_various_kinds() -> std::io::Result<()> {
-        let txns = vec![vec![
-            (0, 0, 1, 1), // native asset transfer
-            (1, 0, 1, 1), // non-native asset transfer with change output
-            (1, 0, 1, 2), // non-native asset transfer with exact change
-        ]];
-        let native_grant = (0, 0, 1);
-        let non_native_grant = (1, 0, 3);
-        test_multixfr_wallet(txns, 2, 1, native_grant, vec![non_native_grant]).await;
-        Ok(())
-    }
-
-    struct MultiXfrParams {
-        max_txns: usize,
-        max_blocks: usize,
-        max_keys: u8,
-        max_defs: u8,
-        max_amt: u64,
-        max_recs: usize,
-    }
-
-    impl MultiXfrParams {
-        const fn new(txns: usize, max_amt: u64) -> Self {
-            // divide txns into 5 blocks
-            let max_txns = if txns > 5 { txns / 5 } else { 1 };
-            let max_blocks = if txns > 5 { 5 } else { txns };
-            // fewer users than txns so we get multiple txns with same key
-            let max_keys = (txns / 2 + 2) as u8;
-            // fewer defs than txns so we get multiple txns with same def
-            let max_defs = (txns / 2 + 1) as u8;
-            // enough records to give everyone 1 of each type, on average
-            // Reasoning for /4:
-            //      E[nkeys] = max_keys/2
-            //      E[ndefs] = max_defs/2
-            // So
-            //      E[nkeys*ndefs] = max_keys*max_defs/4
-            let max_recs = max_keys as usize * max_defs as usize / 4;
-
-            MultiXfrParams {
-                max_txns,
-                max_blocks,
-                max_keys,
-                max_defs,
-                max_amt,
-                max_recs,
-            }
-        }
-
-        fn def(&self) -> impl Strategy<Value = u8> {
-            // range is inclusive because def 0 is the native asset, and other asset defs are
-            // 1-indexed
-            0..=self.max_defs
-        }
-
-        fn key(&self) -> impl Strategy<Value = u8> {
-            0..self.max_keys
-        }
-
-        fn txn_amt(&self) -> impl Strategy<Value = u64> {
-            // Transaction amounts are smaller than record amounts because we don't want to burn a
-            // whole record in one transaction.
-            1..=std::cmp::max(self.max_amt / 5, 2)
-        }
-
-        fn amt(&self) -> impl Strategy<Value = u64> {
-            1..=self.max_amt
-        }
-
-        fn txs(&self) -> impl Strategy<Value = Vec<Vec<(u8, u8, u8, u64)>>> {
-            vec(
-                vec(
-                    (self.def(), self.key(), self.key(), self.txn_amt()),
-                    self.max_txns,
-                ),
-                self.max_blocks,
-            )
-        }
-
-        fn nkeys(&self) -> impl Strategy<Value = u8> {
-            2..=self.max_keys
-        }
-
-        fn ndefs(&self) -> impl Strategy<Value = u8> {
-            1..=self.max_defs
-        }
-
-        fn rec(&self) -> impl Strategy<Value = (u8, u8, u64)> {
-            (self.def(), self.key(), self.amt())
-        }
-
-        fn recs(&self) -> impl Strategy<Value = Vec<(u8, u8, u64)>> {
-            vec(self.rec(), self.max_recs)
-        }
-    }
-
-    const MULTI_XFR_SMALL: MultiXfrParams = MultiXfrParams::new(5, 1000);
-    const MULTI_XFR_LARGE: MultiXfrParams = MultiXfrParams::new(50, 1000);
-
-    proptest! {
-        #![proptest_config(proptest::test_runner::Config {
-            cases: 1,
-            .. proptest::test_runner::Config::default()
-        })]
-
-        #[test]
-        fn proptest_multixfr_wallet_small(
-            txs in MULTI_XFR_SMALL.txs(),
-            nkeys in MULTI_XFR_SMALL.nkeys(),
-            ndefs in MULTI_XFR_SMALL.ndefs(),
-            init_rec in MULTI_XFR_SMALL.rec(),
-            init_recs in MULTI_XFR_SMALL.recs(),
-        ) {
-            block_on(test_multixfr_wallet(txs, nkeys, ndefs, init_rec, init_recs));
-        }
-    }
-
-    proptest! {
-        #![proptest_config(proptest::test_runner::Config {
-            cases: 10,
-            .. proptest::test_runner::Config::default()
-        })]
-
-        #[test]
-        #[ignore]
-        fn proptest_multixfr_wallet_many_small_tests(
-            txs in MULTI_XFR_SMALL.txs(),
-            nkeys in MULTI_XFR_SMALL.nkeys(),
-            ndefs in MULTI_XFR_SMALL.ndefs(),
-            init_rec in MULTI_XFR_SMALL.rec(),
-            init_recs in MULTI_XFR_SMALL.recs(),
-        ) {
-            block_on(test_multixfr_wallet(txs, nkeys, ndefs, init_rec, init_recs));
-        }
-    }
-
-    proptest! {
-        #![proptest_config(proptest::test_runner::Config {
-            cases: 1,
-            .. proptest::test_runner::Config::default()
-        })]
-
-        #[test]
-        #[ignore]
-        fn proptest_multixfr_wallet_one_big_test(
-            txs in MULTI_XFR_LARGE.txs(),
-            nkeys in MULTI_XFR_LARGE.nkeys(),
-            ndefs in MULTI_XFR_LARGE.ndefs(),
-            init_rec in MULTI_XFR_LARGE.rec(),
-            init_recs in MULTI_XFR_LARGE.recs(),
-        ) {
-            block_on(test_multixfr_wallet(txs, nkeys, ndefs, init_rec, init_recs));
-        }
-    }
-}
